% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.controls.r
\name{cases.controls}
\alias{cases.controls}
\title{Generating case control samples}
\usage{
cases.controls(
  nb.cases,
  nb.controls,
  haplos,
  submap,
  beta,
  proba.haplos,
  proba.demes,
  offset.demes,
  tile.length = 20,
  prevalence = 0.01,
  h2 = 0.5,
  kinship = FALSE,
  fraternity = FALSE,
  population = haplos@ped$population,
  ...
)
}
\arguments{
\item{nb.cases}{number of (unrelated) cases}

\item{nb.controls}{number of (unrelated) controls}

\item{haplos}{haplotype bed matrix}

\item{submap}{indices of SNPs with an effect}

\item{beta}{effect}

\item{proba.haplos}{matrix of haplotypes probabilities}

\item{proba.demes}{vector of demes probabilities / weights}

\item{offset.demes}{vector of offset for the environmental component of liability}

\item{tile.length}{tile length in cM}

\item{prevalence}{prevalence}

\item{h2}{heritability}

\item{kinship}{Logical. TRUE to get kinship matrix computed from IBD sharing.}

\item{fraternity}{Logical. TRUE to get fraternity matrix computed from IBD sharing.}

\item{population}{vector giving the population of origin of each haplotype}

\item{...}{addditional parameters give the probability of drawing a tile in each population.}
}
\value{
a list width components `bed`, `kinship` and `fraternity` (if applicable).
}
\description{
Generating case control samples
}
\details{
If provided, `proba.haplos` should be a matrix with as many rows than there are haplotypes in `haplos`. 
If there are several columns in `proba.haplos`, each column corresponds to a deme. The relative weights of the demes
in the total population is specified by `proba.demes`; if missing, all demes have same weight.

If `proba.haplos` is missing, the `population` parameter and the additionnal 
parameters can be used to construct a probabiity matrix. Otherwise, all haplotypes are used with the same probability.

The genotype effect `G.raw` is rescaled in `G.adj` to have variance `h2`. The environmental component `E` has variance `1 - h2`.
Both `G.adj` and `E` are centered unless `offset.demes` values are provided; in that case the values provided give the mean environmental
component in each deme, and `G.adj` is offsetted accordingly. The resulting liability `G.adj + E` is always centered with 
unit variance.
}
\examples{
#' # installs KGH if not already installed
if(!require("KGH")) install.packages("KGH", repos="https://genostats.github.io/R/")
# loads a bed matrix of 1006 european haplotypes
filepath <- system.file("extdata", "1KG_haplos.bed", package = "KGH")
H <- read.bed.matrix(filepath)

# Set of causal SNPs / effects
set.seed(1)
I <- sample.int( ncol(H), 10000 )
beta <- rnorm(length(I))/sqrt(length(I))

# first test. Prevalence 0, generate only controls : general population
# 1000 individuals in 11 demes with different proportions of TSI / IBS
p <- seq(0, 1, length = 11)
x.1 <- cases.controls(0, 1000, H, I, beta, prevalence = 0, TSI = p, IBS = 1 - p)
table(x.1$bed@ped$Deme)
z.1 <- LD.thin(select.snps(x.1$bed, maf > 0.05), 0.1)
K.1 <- GRM(z.1)
plot( eigen(K.1)$vectors, col = hsv(x.1$bed@ped$TSI) )

# second test. Default prevalence and heritability, all haplotypes have equal probabilities.
x.2 <- cases.controls(500, 500, H, I, beta)

# third test. Same with 11 demes with different proportions of TSI / IBS
x.3 <- cases.controls(500, 500, H, I, beta, TSI = p, IBS = 1 - p)

# fourth test. Same with an offset of liability in the demes 
# (environmental liability increases with proportion of TSI)
x.4 <- cases.controls(500, 500, H, I, beta, offset.demes = seq(0,1,length=11), TSI = p, IBS = 1 - p)
with(x.4$bed@ped, table(pheno, TSI))  # more cases with high TSI ancestry
boxplot(x.4$E ~ x.4$Deme)             # liability in the 11 Demes
z.4 <- LD.thin(select.snps(x.4$bed, maf > 0.05), 0.1)
K.4 <- GRM(z.4)
plot( eigen(K.4)$vectors, col = hsv(x.4$bed@ped$TSI), pch = 16+x.4$bed@ped$pheno)
}
