% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.genotypes.r
\name{drop.genotypes}
\alias{drop.genotypes}
\title{Kinship Matrix}
\usage{
drop.genotypes(x, haplos)
}
\arguments{
\item{x}{a list of 'zygotes'}

\item{haplos}{a bed.matrix of haplotypes}
}
\value{
A bed.matrix
}
\description{
Kinship Matrix
}
\details{
If `x` is a list zygotes, returns a bed.matrix 
constructed from the template haplotypes in `haplos`
}
\examples{
#' # installs KGH if not already installed
if(!require("KGH")) install.packages("KGH", repos="https://genostats.github.io/R/")
# loads a bed matrix of 1006 european haplotypes
filepath <- system.file("extdata", "1KG_haplos.bed", package = "KGH")
H <- read.bed.matrix(filepath)

Z <- vector("list", 100)
for(i in 1:100) Z[[i]] <- zygote(1006)

bm <- drop.genotypes(Z, H)
}
