% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.genotypes.vcf.r
\name{drop.genotypes.vcf}
\alias{drop.genotypes.vcf}
\title{Kinship Matrix}
\usage{
drop.genotypes.vcf(x, haplos, filename)
}
\arguments{
\item{x}{a list of 'zygotes'}

\item{haplos}{a bed.matrix of haplotypes}

\item{filename}{file to be created}
}
\value{
NULL object
}
\description{
Kinship Matrix
}
\details{
If `x` is a list zygotes, creates a VCF file (with phased haplotypes)
}
\examples{
#' # installs KGH if not already installed
if(!require("KGH")) install.packages("KGH", repos="https://genostats.github.io/R/")
# loads a bed matrix of 1006 european haplotypes
filepath <- system.file("extdata", "1KG_haplos.bed", package = "KGH")
H <- read.bed.matrix(filepath)

x <- make.inds(100, 1006)
filename <- tempfile(fileext = ".vcf")
drop.genotypes.vcf(x, H, filename)
}
