% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.pop.r
\name{forward}
\alias{forward}
\title{Forward time}
\usage{
forward(zygotes, n.gen, n.keep, lambda = 2)
}
\arguments{
\item{n.gen}{number of generations to simulate}

\item{n.keep}{number of generations to keep}

\item{lambda}{mean number of offspring by pair (default is 2)}

\item{x}{a list of zygotes}
}
\description{
Forward time
}
\examples{
# installs KGH if not already installed
if(!require("KGH")) install.packages("KGH", repos="https://genostats.github.io/R/")
# loads a bed matrix of 1006 european haplotypes
filepath <- system.file("extdata", "1KG_haplos.bed", package = "KGH")
H <- read.bed.matrix(filepath)
# generates 100 individuals
set.seed(1)
x <- make.inds(100, nrow(H))
# forward 10 generations, keep only last two generations
fo <- forward(x, 10, 2)
# Lots of related individuals !
K <- kinship.matrix(fo$zygotes)
round(K[1:10, 1:10], 1)
# to get a bed matrix, use drop.genotypes (here on SNPs 1 to 100)
ybed <- drop.genotypes(fo$zygotes, H[,1:100])
# use father / mother / sex information sent by the function
ybed@ped$father <- fo$father
ybed@ped$mother <- fo$mother
ybed@ped$sex <- fo$sex
}
