% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.drop.tile.index.r
\name{haplotype.drop.tile.index}
\alias{haplotype.drop.tile.index}
\title{Get index of the tile at given positions}
\usage{
haplotype.drop.tile.index(x, chr, pos)
}
\arguments{
\item{x}{a single haplotype}

\item{chr}{vector of chromomose numbers}

\item{pos}{vector of position (in cM)}
}
\value{
an integer vector
}
\description{
Get index of the tile at given positions
}
\details{
sends the tile indexes at the position given in chr, pos.
}
