% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.group.by.tile.index.r
\name{haplotype.group.by.tile.index}
\alias{haplotype.group.by.tile.index}
\title{Group positions sharing the same tile index}
\usage{
haplotype.group.by.tile.index(x, chr, pos)
}
\arguments{
\item{x}{a single haplotype}

\item{chr}{vector of chromomose numbers}

\item{pos}{vector of position (in cM)}
}
\value{
a list
}
\description{
Group positions sharing the same tile index
}
\details{
sends a list giving the indexes of positions sharing the same tiles.
}
