% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.peek.r
\name{haplotype.peek}
\alias{haplotype.peek}
\title{Show haplotype composition}
\usage{
haplotype.peek(x)
}
\value{
A list of data frames.
}
\description{
Show haplotype composition
}
\details{
sends a list of data frames, each with three column: `chr`, `pos`, and `tile`.
The position is in cM. Tile is the tile number (lowest possible value = 1).
}
