% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.poke.r
\name{haplotype.poke}
\alias{haplotype.poke}
\title{Modify a tile number}
\usage{
haplotype.poke(x, tile.index, tile)
}
\arguments{
\item{x}{a single haplotype}

\item{tile.index}{a vector of indexes of tiles to modify}

\item{tile}{new tile numbers}
}
\description{
Modify a tile number
}
\details{
modify (in place) a tile in an haplotype. Tiles are R-indexed (started at 1).
}
