% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.r
\name{haplotype}
\alias{haplotype}
\alias{haplotype.probs}
\title{Mosaic Haplotypes}
\usage{
haplotype(ntiles, mean.length.tiles = 20)

haplotype.probs(probaTiles, mean.length.tiles = 20)
}
\arguments{
\item{ntiles}{number of different tiles}

\item{mean.length.tiles}{mean length of the tiles in cM}

\item{probaTiles}{vector of probabilities for each tiles}
}
\value{
an external pointer to an haplotype
}
\description{
Mosaic Haplotypes
}
\details{
\code{haplotype} generates a mosaic haplotype with tiles numbered from \code{0} to \code{(ntiles - 1)}. 
The haplotype number is drawn uniformly. \code{haplotype.probs} generates a mosaic haplotype
with tiles numbered from \code{0} to \code{length(probaTiles) - 1}, drawn according to the probabilities
in \code{probaTiles}. In both cases, the length of tiles is drawn in a exponential distribution with 
mean length \code{mean.length.tiles}.
}
\seealso{
\link{zygote}, \link{haplotype.peek}
}
