% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBD.segments.r
\name{HBD.segments}
\alias{HBD.segments}
\title{HBD segments}
\usage{
HBD.segments(x)
}
\arguments{
\item{x}{a list of 'zygotes'}
}
\value{
A list of data frames.
}
\description{
HBD segments
}
\details{
If `x` is a list n zygotes, this function compute for each individual
the list of HBD segments along its genome.
}
\examples{
set.seed(1)
x <- zygote(5, 100) # only few founder haplotypes, with long tiles
HBD.segments(x)
}
