% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship.matrix.r
\name{kinship.matrix}
\alias{kinship.matrix}
\title{Kinship Matrix}
\usage{
kinship.matrix(x)
}
\arguments{
\item{x}{a list of 'zygotes'}
}
\value{
A \eqn{n \times n}{n x n} matrix
}
\description{
Kinship Matrix
}
\details{
If `x` is a list n zygotes, this function computes a kinship matrix 
based on the IBD sharing between haplotypes. The coefficients of the matrix
are always non-negative, and the diagonal coefficients are always greater than
or equal to one.
}
\examples{
x1 <- zygote(100)
x2 <- zygote(100)
kinship.matrix( c(x1, x2) )
}
