% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.haps.r
\name{make.haps}
\alias{make.haps}
\title{Generating haplotypes}
\usage{
make.haps(nb.haps, haplos, proba.haplos, tile.length = 20, ibd = FALSE)
}
\arguments{
\item{nb.haps}{number of mosaic haplotypes to generate}

\item{haplos}{(template) haplotype bed matrix}

\item{proba.haplos}{matrix of haplotypes probabilities}

\item{tile.length}{tile length in cM}

\item{ibd}{Logical. TRUE to get a matrix giving the proportion genome shared IBD by two haplotypes.}
}
\value{
a list width components `bed`, and `ibd` (if applicable).
}
\description{
Generating haplotypes
}
\details{
If `proba.haplos` is missing, all haplotypes from `haplos` are used with the same probability.
`proba.haplos` should be a matrix with as many rows than there are haplotypes. If there are several columns, 
in `proba.haplos`, 
`nb.haps` new haplotypes will be generated for each column of probabilities. If `nb.haps` is 
a vector of length `ncol(proba.haplos)`, it specifies the number of haplotypes to be 
generated for each probability set.
}
\examples{
#' # installs KGH if not already installed
if(!require("KGH")) install.packages("KGH", repos="https://genostats.github.io/R/")
# loads a bed matrix of 1006 european haplotypes
filepath <- system.file("extdata", "1KG_haplos.bed", package = "KGH")
H <- read.bed.matrix(filepath)
# Generates a bed matrix with 1000 haplotypes
x <- make.haps(1000, H)
}
