% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.inbreds.r
\name{make.inbreds}
\alias{make.inbreds}
\title{Generating unrelated individuals with fiexed inbreding coeffs}
\usage{
make.inbreds(
  nb.inds,
  ntiles,
  f,
  a,
  hbd.length,
  non.hbd.length,
  tile.length = 20,
  segments = FALSE
)
}
\arguments{
\item{nb.inds}{number of (unrelated) individuals}

\item{ntiles}{number of distinct tiles}

\item{f}{inbreding coefficient(s)}

\item{a}{parameter for the length of segments in HMM (in \eqn{cM^{-1}})}

\item{hbd.length}{mean length of HBD segments in cM}

\item{non.hbd.length}{mean length of non HBD segments in cM}

\item{tile.length}{tile length in cM}

\item{segments}{Logical. TRUE to get the position of HBD segments}
}
\value{
a list with components `bed.matrix`, `inbred.coef` (computed from the HBD length) 
and `segments` (if applicable).
}
\description{
Generating unrelated individuals with fiexed inbreding coeffs
}
\details{
Specify a and f or hbd.length and non.hbd.length, but not both.
This function uses a slightly imperfect method to generate individuals with a given HBD:
Haplotypes H1 and A are generated with tiles of mean length \code{tile.length}, then H2
is generated by taking tiles from H1 and A with mean lengthes \code{hbd.length} and \code{non.hbd.length}.

The HBD segments length are taken in an exponential distribution. If the parameters \code{hbd.length} and 
\code{non.hbd.length} are not specified, the mean HBD length is \eqn{\frac{1}{a(1-f)}}{1/(a(1-f))}
and the mean non-HBD length is \eqn{\frac{1}{af}}{1/(af)}.

If \code{segments = TRUE} then a list of data frame giving the position of HBD segments (in cM) is returned.
}
\examples{
#' # installs KGH if not already installed
if(!require("KGH")) install.packages("KGH", repos="https://genostats.github.io/R/")
# loads a bed matrix of 1006 european haplotypes
filepath <- system.file("extdata", "1KG_haplos.bed", package = "KGH")
H <- read.bed.matrix(filepath)
# Generates 100 individuals (zygotes) with specified a and f values
set.seed(1)
x <- make.inbreds(100, nrow(H), f = 0.02, a = 0.05, segments = TRUE)
# their "true" indbreeding coefficients 
mean(x$inbred.coef)
hist(x$inbred.coef)
# their HBD segments
x$segments[[1]]
# to get a bed matrix, use drop.genotypes (here on SNPs 1 to 100)
xbed <- drop.genotypes(x$zygotes, H[,1:100])
}
