% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.inds.r
\name{make.inds}
\alias{make.inds}
\title{Generating unrelated individuals with tile haplotypes}
\usage{
make.inds(
  nb.inds,
  haplos,
  proba.haplos,
  tile.length = 20,
  kinship = FALSE,
  fraternity = FALSE
)
}
\arguments{
\item{nb.inds}{number of (unrelated) individuals}

\item{haplos}{haplotype bed matrix}

\item{proba.haplos}{matrix of haplotypes probabilities}

\item{tile.length}{tile length in cM}

\item{kinship}{Logical. TRUE to get kinship matrix computed from IBD sharing.}

\item{fraternity}{Logical. TRUE to get fraternity matrix computed from IBD sharing.}
}
\value{
a list width components `bed`, `kinship` and `fraternity` (if applicable).
}
\description{
Generating unrelated individuals with tile haplotypes
}
\details{
If `proba.haplos` is missing, all haplotypes are used with the same probability.
`proba.haplos` should be a matrix with as many rows of haplotypes. If there are several columns, 
`nb.inds` individuals will be generated for each column of probabilities.
}
\examples{
#' # installs KGH is not already installed
if(!require("KGH")) install.packages("KGH", repos="https://genostats.github.io/R/")
# loads a bed matrix of 1006 european haplotypes
filepath <- system.file("extdata", "1KG_haplos.bed", package = "KGH")
H <- read.bed.matrix("~/PROJECTS/PCA/2019/1KG_haplos")
# Generates a bed matrix with 1000 individuals
x <- make.inds(1000, H)
}
