% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.r
\name{peek}
\alias{peek}
\alias{haplotype_peek}
\alias{zygote_peek}
\title{Peek into haplotypes and zygotes}
\usage{
haplotype_peek(x)

zygote_peek(x)
}
\arguments{
\item{x}{a haplotype or a zygote}
}
\description{
Peek into haplotypes and zygotes
}
\details{
\code{haplotype_peek} sends back a data frame with columns 'chr', 'pos', 'tile'.
The position in 'pos' is the endpoint of the tile (in cM) whose number is in 'tile'. 
\code{zygote_peek} sends back a list of two such data frames.
}
\examples{
# an example with long tiles
set.seed(1)
x <- haplotype(10, 100)
haplotype_peek(x)
}
