% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.r
\name{reproduce}
\alias{reproduce}
\title{Mating}
\usage{
reproduce(z1, z2)
}
\arguments{
\item{z1, z2}{zygote vectors of same length}
}
\value{
A zygote vector
}
\description{
Mating
}
\details{
This function will create offsprings of individuals in z1 x z2
}
\examples{
fa <- make.inds(2, 1000)
mo <- make.inds(2, 1000)
of <- reproduce(fa, mo)
K <- kinship.matrix( c(fa, mo, of) )
round(K, 1)
# You can also use `+` for a simple call. 
of2 <- fa + mo
L <- kinship.matrix( c(of, of2) )
round(L, 1)
}
