% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.gen.r
\name{vcf.gen}
\alias{vcf.gen}
\title{Generates VCF file}
\usage{
vcf.gen(
  x,
  filename,
  depth1,
  depth2,
  eps = 10^((-0.1) * runif(ncol(x), 20, 30)),
  subset
)
}
\arguments{
\item{x}{}

\item{filename}{}

\item{depth1}{}

\item{depth2}{}

\item{eps}{}
}
\value{

}
\description{
Generates VCF file
}
\examples{
x <- as.bed.matrix(LCT.gen, LCT.fam, LCT.bim)
vcf.gen(x, "/tmp/LCT.vcf", 5, 10)
}
