% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygote.drop.tile.index.r
\name{zygote.drop.tile.index}
\alias{zygote.drop.tile.index}
\title{Get index of the tile at given positions}
\usage{
zygote.drop.tile.index(x, chr, pos)
}
\arguments{
\item{x}{a single zygote}

\item{chr}{vector of chromomose numbers}

\item{pos}{vector of position (in cM)}
}
\value{
a list of two integer vectors, for the two haplotypes
}
\description{
Get index of the tile at given positions
}
\details{
sends the tile indexes at the position given in chr, pos.
}
