% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygote.group.by.tile.index.r
\name{zygote.group.by.tile.index}
\alias{zygote.group.by.tile.index}
\title{Group positions sharing the same tile index}
\usage{
zygote.group.by.tile.index(x, chr, pos)
}
\arguments{
\item{x}{a single zygote}

\item{chr}{vector of chromomose numbers}

\item{pos}{vector of position (in cM)}
}
\value{
a list
}
\description{
Group positions sharing the same tile index
}
\details{
sends two list giving the indexes of positions sharing the same tiles, for the two haplotypes
}
