% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygote.peek.r
\name{zygote.peek}
\alias{zygote.peek}
\title{Show zygote composition}
\usage{
zygote.peek(x)
}
\value{
A list of data frames.
}
\description{
Show zygote composition
}
\details{
sends a list of data frames, each with three column: `chr`, `pos`, and `tile`.
The position is in cM. Tile is the tile number (lowest possible value = 1).
}
