% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygote.poke.r
\name{zygote.poke}
\alias{zygote.poke}
\title{Modify a tile number}
\usage{
zygote.poke(x, tile.index.a, tile.a, tile.index.b, tile.b)
}
\arguments{
\item{x}{a single zygote}

\item{tile.index.a}{a vector of index of tiles to modify on first haplotype}

\item{tile.a}{a vector of new tile numbers}

\item{tile.index.b}{a vector of index of tiles to modify on second haplotype}

\item{tile.b}{a vector of new tile numbers}
}
\description{
Modify a tile number
}
\details{
modify (in place) tiles in a zygote. Tiles are R-indexed.
}
