% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.genotypes.r
\name{drop.genotypes}
\alias{drop.genotypes}
\title{Kinship Matrix}
\usage{
drop.genotypes(x, haplos, phased = FALSE)
}
\arguments{
\item{x}{a list of 'zygotes'}

\item{haplos}{a bed.matrix of haplotypes}

\item{phased}{a boolean}
}
\value{
A bed.matrix
}
\description{
Kinship Matrix
}
\details{
If `x` is a list zygotes, returns a bed.matrix 
constructed from the template haplotypes in `haplos`. If phased is TRUE,
instead of genotypes, two haplotypes are built for each individual.
}
\examples{
#' # installs KGH if not already installed
if(!require("KGH")) install.packages("KGH", repos="https://genostats.github.io/R/")
# loads a bed matrix of 1006 european haplotypes
filepath <- system.file("extdata", "1KG_haplos.bed", package = "KGH")
H <- read.bed.matrix(filepath)

x <- make.inds(100, 1006)

bm <- drop.genotypes(x, H)
}
