% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.r
\name{reproduce}
\alias{reproduce}
\title{Mating}
\arguments{
\item{z1, z2}{zygotes}
}
\value{
A zygote
}
\description{
Mating
}
\details{
If `z1` and `z2` are zyogtes, this function will proceed to
meiosis and send back an offspring.
}
\examples{
z1 <- zygote(100)
z2 <- zygote(100)
z3 <- reproduce(z1, z2)
kinship.matrix( c(z1, z2, z3) )

}
