% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygote.r
\name{zygote}
\alias{zygote}
\alias{zygote_prob}
\title{Zygotes}
\usage{
zygote(ntiles, mean_length_tiles = 20)

zygote_probs(probaTiles, mean_length_tiles = 20)
}
\arguments{
\item{ntiles}{number of different tiles}

\item{probaTiles}{vector of probabilities for each tiles}

\item{mean_length_tiles}{mean length of the tiles in cM}
}
\value{
an external pointer to an haplotype
}
\description{
Zygotes
}
\details{
\code{haplotype} generates a mosaic haplotype with tiles numbered from \code{0} to \code{(ntiles - 1)}. 
The haplotype number is drawn uniformly. \code{haplotype_probs} generates a mosaic haplotype
with tiles numbered from \code{0} to \code{length(probaTiles) - 1}, drawn according to the probabilities
in \code{probaTiles}. In both cases, the length of tiles is drawn in a exponential distribution with 
mean length \code{mean_length_tiles}.
}
\seealso{
\link{haplotype}, \link{zygote_peek}
}
